server/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.13.0  
openpyxl